+++ docs/changelog.md
@@ -9,6 +9,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 
 # Unreleased
 
+- Fail when attribute is mispelled.
 - CentOS 8 support.
 - PostgreSQL 13 support.
 - Fix join order.